fix(nn): round fused products before the add in KAN, MemorizingAttention and MTA
Same defect as the TPA fix in the previous commit, in the three other fused
paths that the recent merge batch introduced. Each is written `acc += a * b`,
which the Go compiler CONTRACTS into a fused multiply-add on arm64 and does not
on amd64. An FMA rounds once where a separate multiply and add round twice, so
each fused path computes a different number per architecture and its bit-exact
pin against the dispatch path fails on arm64 while amd64 CI stays green.

  KAN                 fused vs einsum differed by 1.11e-16
  MemorizingAttention 0.2515599025656111 vs 0.251559902565611
  MTA headConv        0.46783668 vs 0.4678367 (the F32 arm; F64 fixed too)

nn failures: 10 -> 6. The four the merge batch introduced are closed.

KAN is worth singling out: its doc comment already claimed the fused path used
"no FMA", and on arm64 that was simply untrue. The comment now says how the
property is actually obtained rather than asserting it.

MTA needed both dtype arms — the F32 one is where the test happened to fail
first, but the F64 arm has the identical shape and would diverge on the same
hardware.

THE REMAINING SIX PREDATE THIS: DeltaNet, EMAUpdate, GLA, HGRNSeq, RGLRUSeq and
TitansLinear were already failing before the merge batch, are untouched here,
and need their own diagnosis — a blanket wrap of every mul-add across
deltanet/gla/titans was tried and fixed none of them, so their divergence is
something else.

+++ nn/memorizing_attention.go
@@ -297,7 +297,9 @@ func memScoresFusedF64(qh, kg *tensor.Tensor, t, topM, dk int) *tensor.Tensor {
 			kb := kBase + mm*dk
 			var s float64
 			for d := 0; d < dk; d++ {
-				s += qs[qBase+d] * ks[kb+d]
+				// rounded before the add: a bare mul-add contracts to FMA on arm64 only, which
+				// is what broke this path's bit-exact pin against dispatch while amd64 CI passed.
+				s += float64(qs[qBase+d] * ks[kb+d])
 			}
 			os[oBase+mm] = s
 		}
@@ -321,7 +323,7 @@ func memOutFusedF64(prob, vg *tensor.Tensor, t, topM, dk int) *tensor.Tensor {
 			p := ps[pBase+mm]
 			vb := vBase + mm*dk
 			for d := 0; d < dk; d++ {
-				os[oBase+d] += p * vs[vb+d]
+				os[oBase+d] += float64(p * vs[vb+d])
 			}
 		}
 	}
